workshop adv part 2

+++ JS courses/ReactJS/07. ReactJS-Routing-Workshop2 2/games-play/src/components/GameDetails/GameDetails.js
@@ -1,47 +1,60 @@
-import { useEffect, useState, useContext } from 'react';
+import { useEffect, useState, useReducer } from 'react';
 import { useParams, useNavigate, Link } from 'react-router-dom';
 
 import { gameServiceFactory } from '../../services/gameService';
+import * as commentService from '../../services/commentService';
 import { useService } from '../../hooks/useService';
-import { AuthContext } from '../../contexts/AuthContext';
+import { useAuthContext } from '../../contexts/AuthContext';
+
+import { AddComment } from './AddComment/AddComment';
+import { gameReducer } from '../../reducers/gameReducer';
+import { useGameContext } from '../../contexts/GameContext';
 
 export const GameDetails = () => {
-    const { userId } = useContext(AuthContext);
-    const [username, setUsername] = useState('');
-    const [comment, setComment] = useState('');
     const { gameId } = useParams();
-    const [game, setGame] = useState({});
+    const { userId, isAuthenticated, userEmail } = useAuthContext();
+    const { deleteGame } = useGameContext();
+    const [game, dispatch] = useReducer(gameReducer, {});
     const gameService = useService(gameServiceFactory)
     const navigate = useNavigate();
 
     useEffect(() => {
-        gameService.getOne(gameId)
-            .then(result => {
-                setGame(result);
-            })
+        Promise.all([
+            gameService.getOne(gameId),
+            commentService.getAll(gameId),
+        ]).then(([gameData, comments]) => {
+            const gameState = {
+                ...gameData,
+                comments,
+            };
+            
+            dispatch({type: 'GAME_FETCH', payload: gameState})
+        });
     }, [gameId]);
 
-    const onCommentSubmit = async (e) => {
-        e.preventDefault();
+    const onCommentSubmit = async (values) => {
+        const response = await commentService.create(gameId, values.comment);
 
-        const result = await gameService.addComment(gameId, {
-            username,
-            comment,
+        dispatch({
+            type: 'COMMENT_ADD',
+            payload: response,
+            userEmail,
         });
-
-        setGame(state => ({ ...state, comments: { ...state.comments, [result._id]: result } }));
-        setUsername('');
-        setComment('');
     };
 
     const isOwner = game._ownerId === userId;
 
     const onDeleteClick = async () => {
-        await gameService.delete(game._id);
+        // eslint-disable-next-line no-restricted-globals
+        const result = confirm(`Are you sure you want to delete ${game.title}`);
 
-        // TODO: delete from state
+        if (result) {
+            await gameService.delete(game._id);
 
-        navigate('/catalog');
+            deleteGame(game._id);
+
+            navigate('/catalog');
+        }
     };
 
     return (
@@ -61,19 +74,18 @@ export const GameDetails = () => {
                 <div className="details-comments">
                     <h2>Comments:</h2>
                     <ul>
-                        {game.comments && Object.values(game.comments).map(x => (
+                        {game.comments && game.comments.map(x => (
                             <li key={x._id} className="comment">
-                                <p>{x.username}: {x.comment}</p>
+                                <p>{x.author.email}: {x.comment}</p>
                             </li>
                         ))}
                     </ul>
 
-                    {/* {!Object.values(game.comments).length && (
+                    {!game.comments?.length && (
                         <p className="no-comment">No comments.</p>
-                    )} */}
+                    )}
                 </div>
 
-                {/* <!-- Edit/Delete buttons ( Only for creator of this game )  --> */}
                 {isOwner && (
                     <div className="buttons">
                         <Link to={`/catalog/${game._id}/edit`} className="button">Edit</Link>
@@ -82,17 +94,7 @@ export const GameDetails = () => {
                 )}
             </div>
 
-            {/* <!-- Bonus --> */}
-            {/* <!-- Add Comment ( Only for logged-in users, which is not creators of the current game ) --> */}
-            <article className="create-comment">
-                <label>Add new comment:</label>
-                <form className="form" onSubmit={onCommentSubmit}>
-                    <input type="text" name="username" placeholder='Пешо' value={username} onChange={(e) => setUsername(e.target.value)} />
-                    <textarea name="comment" placeholder="Comment......" value={comment} onChange={(e) => setComment(e.target.value)}></textarea>
-                    <input className="btn submit" type="submit" value="Add Comment" />
-                </form>
-            </article>
-
+            {isAuthenticated && <AddComment onCommentSubmit={onCommentSubmit} />}
         </section>
     );
 };
